+++ datastar/shared/src/main/scala-3/zio/blocks/datastar/SignalVersionSpecific.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package zio.blocks.datastar
+package zio.http.datastar
 
 import scala.quoted.*
