+++ datastar/shared/src/main/scala/zio/blocks/datastar/DataOn.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package zio.blocks.datastar
+package zio.http.datastar
 
-import zio.blocks.html.Dom
+import zio.http.html.Dom
 import zio.blocks.maybe.Maybe
 
 /**
