+++ datastar/shared/src/main/scala/zio/blocks/datastar/DatastarRef.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package zio.blocks.datastar
+package zio.http.datastar
 
-import zio.blocks.html.ToJs
+import zio.http.html.ToJs
 
 final class DatastarRef private[datastar] (val signalName: String) {
   val value: String = "$" + signalName
