+++ datastar/shared/src/main/scala/zio/blocks/datastar/ElementPatchMode.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package zio.blocks.datastar
+package zio.http.datastar
 
 /**
  * Describes how Datastar should patch rendered content into the target element.
