+++ datastar/shared/src/main/scala/zio/blocks/datastar/SignalUpdate.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package zio.blocks.datastar
+package zio.http.datastar
 
-import zio.blocks.html.ToJs
+import zio.http.html.ToJs
 
 /**
  * A signal name paired with its JSON-serialized value.
