+++ datastar/shared/src/main/scala/zio/blocks/datastar/package.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package zio.blocks
+package zio.http
 
 package object datastar extends DatastarAttributes {
