+++ datastar/shared/src/test/scala/zio/blocks/datastar/DatastarAttributesSpec.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package zio.blocks.datastar
+package zio.http.datastar
 
-import zio.blocks.html.{dataAttr => _, _}
+import zio.http.html.{dataAttr => _, _}
 import zio.blocks.schema.Schema
 import zio.test._
 
@@ -79,8 +79,8 @@ object DatastarAttributesSpec extends ZIOSpecDefault {
       },
       test("dataText := raw String does not compile") {
         typeCheck("""
-          import zio.blocks.datastar._
-          import zio.blocks.html._
+          import zio.http.datastar._
+          import zio.http.html._
           div(dataText := "$count")
         """).map { result =>
           assertTrue(result.isLeft) &&
@@ -133,8 +133,8 @@ object DatastarAttributesSpec extends ZIOSpecDefault {
       },
       test("dataOn.click := raw String does not compile") {
         typeCheck("""
-          import zio.blocks.datastar._
-          import zio.blocks.html._
+          import zio.http.datastar._
+          import zio.http.html._
           div(dataOn.click := "$count++")
         """).map { result =>
           assertTrue(result.isLeft) &&
@@ -266,8 +266,8 @@ object DatastarAttributesSpec extends ZIOSpecDefault {
       },
       test("dataInit := raw String does not compile") {
         typeCheck("""
-          import zio.blocks.datastar._
-          import zio.blocks.html._
+          import zio.http.datastar._
+          import zio.http.html._
           div(dataInit := "@get('/data')")
         """).map { result =>
           assertTrue(result.isLeft) &&
